requirements.txt
bayesmark==0.0.8
DBUtils==3.0.3
GPy==1.10.0
gpytorch==1.8.1
gym==0.26.2
matplotlib==3.5.3
numpy==1.21.6
pandas==1.3.5
paramiko==3.2.0
POAP==0.1.26
psqlparse==1.0rc7
psycopg2-binary
pyDOE==0.3.8
pymoo==0.6.0.1
pySOT==0.3.3
pytest==7.3.2
scikit_learn==1.0.2
scipy==1.7.3
setuptools==65.6.3
shap==0.41.0
torch==1.13.1
xgboost==1.6.2
psqlparse==1.0rc7